==> 3-detroit-auctions/code/05-mcnulty-cleaning.ipynb <==
# 0.5 McNulty Cleaning

Loads data, performs tts, and saves train/test dfs as .pkl files.

import mcnultymod
import pandas as pd
import numpy as np
# from sklearn.model_selection import train_test_split
# import matplotlib.pyplot as plt
# from mpl_toolkits.basemap import Basemap
# import seaborn as sns
# plt.style.use('seaborn')
# %matplotlib inline
import pickle

# from datetime import datetime, timedelta, date

auctions_raw = mcnultymod.fetch_data('auctionsclosed', '52.14.0.197')

auctions_raw.shape

auctions, auctions_test = mcnultymod.split_and_clean(auctions_raw)

auctions.purchasertype.value_counts()[0]

for which_set, df in zip(["TRAIN", "TEST"], [auctions, auctions_test]):
    inv = df.purchasertype.value_counts()[1] / df.shape[0]
    hb = df.purchasertype.value_counts()[0] / df.shape[0]
    print(f"\n{which_set}")
    print("Investor ratio:", inv)
    print("Homebuyer ratio:", hb)

auctions.to_pickle('../data/auctionsclosed.pkl')
auctions_test.to_pickle('../data/auctionsclosed_test.pkl')
