==> 3-detroit-auctions/code/10-mcnulty-feature-engineering-final.ipynb <==
import mcnultymod
import pandas as pd
import numpy as np
# import matplotlib.pyplot as plt
# from mpl_toolkits.basemap import Basemap
# import seaborn as sns
# plt.style.use('seaborn')
# %matplotlib inline
# from geopy.distance import vincenty

import pickle

auctions = pd.read_pickle('../data/auctionsclosed.pkl')
auctions_test = pd.read_pickle('../data/auctionsclosed_test.pkl')

Performs feature engineering on both the train and test set. **All transforms are fit on the training set, exactly like with Standard Scaler — the test features never "see" their corresponding targets.** 

auctions_v2, auctions_test_v2 = mcnultymod.feature_eng([auctions, auctions_test])

auctions_v2.to_pickle("../data/auctionsclosed_v2.pkl")

auctions_test_v2.to_pickle("../data/auctionsclosed_test_v2.pkl")
