==> 3-detroit-auctions/code/30-mcnulty-model-dev.ipynb <==
# 3.0 McNulty Model Development

import mcnultymod
import pandas as pd
import numpy as np
import warnings

from sklearn.preprocessing import StandardScaler
from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC, LinearSVC
from sklearn.naive_bayes import GaussianNB
# from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn import metrics
from sklearn.model_selection import (train_test_split,
                                     cross_validate,
                                     cross_val_score,
                                     # cross_val_predict,
                                     learning_curve,
                                     StratifiedKFold,
                                     GridSearchCV,
                                     RandomizedSearchCV
                                     # KFold
                                    )

from datetime import datetime, timedelta, date
# import os
import pickle

import matplotlib.pyplot as plt
from mpl_toolkits.basemap import Basemap
import seaborn as sns
plt.style.use('seaborn')
%matplotlib inline

## 1. First Round of Features

auctions = pd.read_pickle('../data/auctionsclosed.pkl')
auctions_test = pd.read_pickle('../data/auctionsclosed_test.pkl')

# What's the right way to do this...?
auctions['ptype_num'] = 0
# Evidently not this:
# auctions.ptype_num[auctions.purchasertype == 'Investor'] = 1
# Supposedly this is, but it still throws an error message:
# auctions.loc[:, 'ptype_num'][auctions.purchasertype == 'Investor'] = 1
# THIS IS IT! NEVER FORGET:
auctions.loc[auctions.purchasertype == 'Investor', 'ptype_num'] = 1

auctions.head()

auctions.sample(10)

For the inital run, I'll use price (continuous), lat/lon (continuous) and council district (categorical). If I have time before MVP, I'll bin neighborhood in some way, though that would inject colinearity with council district, and I don't think it would be that useful anyway.

auctions_short = auctions.filter(['ptype_num', 'price', 'latitude', 'longitude', 'councildistrict'])
auctions_short.sample(5)

auctions_short.corr()

y = auctions_short.iloc[:, 0]
X = auctions_short.iloc[:, 1:]

Sets certain columns to categories, creates dummies, standardizes data:

X_std = mcnultymod.prep_X(['councildistrict'], X)

### The Baseline: Guessing Investor for Everything

mcnultymod.print_scores('Dummy (guessing Investor)', mcnultymod.the_dummy(y))

### KNN

knn_scores = mcnultymod.try_some_ks(X_std, y, max_k=15)

mcnultymod.print_knn_scores(knn_scores)

### Logistic Regresion, SVM

model_list = [
    LogisticRegression(class_weight='balanced', random_state=23),
    LinearSVC(class_weight='balanced', random_state=23),
    SVC(class_weight='balanced', random_state=23),
    GaussianNB()
]

name_list = [
    'Logistic Regression',
    'Linear Support Vector Classifier',
    'Support Vector Classifier with RBF Kernel',
    'Naive Bayes (Gaussian)'
]

model_scores = mcnultymod.crossval(X_std, y, model_list, name_list)

Our dummy had an F1 score of 0.4569. We're not even close... need more data.

## 2. With A Few More Features

auctions_v2 = pd.read_pickle('../data/auctionsclosed_v2.pkl')

auctions_v2.head()

auctions_v2_short = (auctions_v2
                     .filter(['ptype_num',
                              'p_bin',
                              'n_bin',
                              'thr_0',
                              'tcount_0'
                             ]))

auctions_v2_short.corr()

yv2 = auctions_v2_short.iloc[:, 0]
Xv2 = auctions_v2_short.iloc[:, 1:]

Xv2_std = mcnultymod.prep_X(Xv2.columns[:-1], Xv2)

### LR, SVM, NB

model_scores = mcnultymod.crossval(Xv2_std, yv2, model_list, name_list)

With 10 folds instead of 5:

model_scores = mcnultymod.crossval(Xv2_std, yv2, model_list, name_list, folds=10)

And the dummy, for baseline comparison:

mcnultymod.print_scores('Dummy (guessing Investor)', mcnultymod.the_dummy(yv2))

### RF

Since one can't do CV with Random Forest, here's a DIY tuning parameters based on training and OOB accuracy scores.

rf_scores = mcnultymod.rf_grid(Xv2_std, 
                               yv2, 
                               num_estimators=[15, 20, 50, 100, 500, 1000, 2000, 5000], 
                               max_depths=[3, 2]
                              )

#### Final round will include RF models with 5000 trees/max depth 2.

If time — make dictionary keys just the number of trees and number for max depth.

## 3. Grid Search on LR and SVC (RBF)

If time (or after presentation) — wrap this into a function.

# rand = RandomizedSearchCV(knn, param_dist, cv=10, scoring='accuracy', n_iter=10, random_state=42)

# LogisticRegression(class_weight='balanced')
# LinearSVC(class_weight='balanced')

### LR

lr_pgrid = {
    'penalty': ['l1', 'l2'],
    'C': [i / 10 for i in range(1,11)]
}

# lr_grid = RandomizedSearchCV(
#     LogisticRegression(class_weight='balanced'),
#     lr_pgrid,
#     cv=10,
#     scoring='f1'
#     n_iter=10,
#     random_state=23
# )

Because there are only 20 combinations to investigate, normal GridSearchCV will suffice.

lr_grid = GridSearchCV(
    LogisticRegression(class_weight='balanced', random_state=23),
    lr_pgrid,
    cv=10,
    scoring='f1'
)

lr_grid.fit(Xv2_std, yv2)

lr_grid.grid_scores_

print(lr_grid.best_score_)
print(lr_grid.best_params_)
print(lr_grid.best_estimator_)

### SVC Linear

svc_pgrid = {
    # 'penalty': ['l1', 'l2'],
    # 'loss': ['hinge', 'squared_hinge'],
    'C': [i / 10 for i in range(1,11)]
}

# svc_grid = RandomizedSearchCV(
#     LinearSVC(class_weight='balanced', random_state=23),
#     svc_pgrid,
#     cv=10,
#     scoring='f1',
#     n_iter=20,
#     random_state=23
# )

svc_grid = GridSearchCV(
    LinearSVC(class_weight='balanced', random_state=23),
    svc_pgrid,
    cv=10,
    scoring='f1'
)

svc_grid.fit(Xv2_std, yv2)

svc_grid.grid_scores_

print(svc_grid.best_score_)
print(svc_grid.best_params_)
print(svc_grid.best_estimator_)
