==> 3-detroit-auctions/code/40-flask-sketch.ipynb <==
import pandas as pd
import numpy as np
# from googlemaps import GoogleMaps
import googlemaps
from geopy.distance import vincenty
import os


If I wanted to push a .png to server, could put it in 'static' folder. Will be tricky. Could render main map when server loads, save the Basemap object, make a deepcopy, then tack another dot onto it and export PNG when needed.

gmaps_api_key = os.environ['gmAPI']

gmaps = googlemaps.Client(key=gmaps_api_key)

address = "1077 Hubbard St, Detroit, MI"
result = gmaps.geocode(address)
lat = result[0]['geometry']['location']['lat']
lng = result[0]['geometry']['location']['lng']

print(f"{lat}, {lng}")

# 0.01449 latitude points per mile (0.015)
vincenty((42.3, -83.1), (42.315, -83.1)).miles

# 0.01952 longitude points per mile (0.02)
vincenty((42.3, -83.1), (42.3, -83.12)).miles
